// site/src/types/types.ts
// Code generated by 'guts'. DO NOT EDIT.

// From web/diagnostic.go
export interface Diagnostic {
    readonly severity: DiagnosticSeverity;
    readonly summary: string;
    readonly detail: string;
}

// From web/diagnostic.go
export type DiagnosticSeverity = "error" | "warning";

export const DiagnosticSeveritys: DiagnosticSeverity[] = ["error", "warning"];

// From web/diagnostic.go
export type Diagnostics = readonly Diagnostic[];

// From types/parameter.go
export interface Parameter extends RichParameter {
    readonly value: string;
}

// From types/parameter.go
export interface ParameterOption {
    readonly name: string;
    readonly description: string;
    readonly value: string;
    readonly icon: string;
}

// From types/enum.go
export type ParameterType = "boolean" | "list(string)" | "number" | "string";

export const ParameterTypes: ParameterType[] = ["boolean", "list(string)", "number", "string"];

// From types/parameter.go
export interface ParameterValidation {
    readonly validation_regex: string | null;
    readonly validation_error: string;
    readonly validation_min: number | null;
    readonly validation_max: number | null;
    readonly validation_monotonic: string | null;
}

// From web/session.go
export interface Request {
    readonly ID: number;
    readonly Inputs: Record<string, string>;
}

// From web/session.go
export interface Response {
    readonly diagnostics: Diagnostics;
    readonly parameters: readonly Parameter[];
}

// From types/parameter.go
export interface RichParameter {
    readonly name: string;
    readonly display_name: string;
    readonly description: string;
    readonly diagnostics: Diagnostics;
    readonly form_control: string;
    readonly type: ParameterType;
    readonly mutable: boolean;
    readonly default_value: string;
    readonly icon: string;
    readonly options: readonly (ParameterOption)[];
    readonly validations: readonly (ParameterValidation)[];
    readonly required: boolean;
    readonly order: number;
    readonly ephemeral: boolean;
}